+++ app/Controllers/Http/GamesController.ts
@@ -12,7 +12,7 @@ export default class GamesController {
 
   public async store({ request, auth }: HttpContextContract) {
     const user = await User.findOrFail(auth.user?.id)
-    const user_level = await UserLevelAccess.findOrFail(user.id)
+    const user_level = await UserLevelAccess.findByOrFail('user_id', user.id)
 
     if(user_level.level_access_id === 1){
       const data = await request.only(['type', 'description', 'range', 'price', 'max_number', 'color'])
@@ -40,34 +40,43 @@ export default class GamesController {
     }
   }
 
-  public async update({ params, request}: HttpContextContract) {
-    try{
-      const data = await request.only(['type', 'description', 'range', 'price', 'max_number', 'color'])
-      const game = await Game.findOrFail(params.id)
+  public async update({ params, request, auth}: HttpContextContract) {
+    const user = await User.findOrFail(auth.user?.id)
+    const user_level = await UserLevelAccess.findByOrFail('user_id', user.id)
+    if(user_level.level_access_id === 1){
       try{
-        game.merge(data)
-        await game.save()
-        return {game: game}
+        const data = await request.only(['type', 'description', 'range', 'price', 'max_number', 'color'])
+        const game = await Game.findOrFail(params.id)
+        try{
+          game.merge(data)
+          await game.save()
+          return {game: game}
+        }catch{
+          return {Error: `Error on update game ${game.id}`}
+        }
       }catch{
-        return {Error: `Error on update game ${game.id}`}
+        return {Error: 'Game not found, or column invalid'}
       }
-    }catch{
-      return {Error: 'Game not found, or column invalid'}
+    }else{
+      return {Error: `You don't have permission to change a game, ask for the administrator.`}
     }
   }
 
-  public async destroy({ params }: HttpContextContract) {
-
-    try{
-      const game = await Game.findOrFail(params.id)
+  public async destroy({ params, auth }: HttpContextContract) {
+    const user = await User.findOrFail(auth.user?.id)
+    const user_level = await UserLevelAccess.findByOrFail('user_id', user.id)
+    if(user_level.level_access_id === 1){
       try{
-        game.delete()
-        return {deleted: true}
+        const game = await Game.findOrFail(params.id)
+        try{
+          game.delete()
+          return {deleted: true}
+        }catch{
+          return {Error: 'Error on delete game'}
+        }
       }catch{
-        return {Error: 'Error on delete game'}
+        return {error: 'Game do not found'}
       }
-    }catch{
-      return {error: 'Game do not found'}
     }
 
   }
